requirements.txt
africastalking==1.1.4
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11
django-cors-headers==2.3.0
django-extensions==2.0.7
django-filter==1.1.0
django-sendgrid==1.0.1
django-toolbelt==0.0.1
djangorestframework==3.8.2
gunicorn==19.7.1
idna==2.7
Markdown==2.6.11
psycopg2==2.7.4
psycopg2-binary==2.7.4
pydot==1.2.4
pyparsing==2.2.0
pytz==2018.5
requests==2.19.1
schema==0.6.8
six==1.11.0
static3==0.7.0
urllib3==1.23
